Create helper file in integration folder
